--- examples/demos/polygon/index.ts
@@ -1,8 +1,0 @@
-export { MapRender as extrude } from './extrude';
-export { MapRender as extrudeCity } from './extrude-city';
-export { MapRender as extrusion } from './extrusion';
-export { MapRender as fill } from './fill';
-export { MapRender as fillLinear } from './fill-linear';
-export { MapRender as ocean } from './ocean';
-export { MapRender as texture } from './texture';
-export { MapRender as water } from './water';
