--- examples/demos/polygon/texture.ts
@@ -1,48 +1,0 @@
-import { PolygonLayer, Scene } from '@antv/l7';
-import * as allMap from '@antv/l7-maps';
-import type { RenderDemoOptions } from '../../types';
-
-export function MapRender(options: RenderDemoOptions) {
-  const scene = new Scene({
-    id: 'map',
-    renderer: options.renderer,
-    map: new allMap[options.map]({
-      style: 'light',
-      center: [121.434765, 31.256735],
-      zoom: 14.83,
-    }),
-  });
-
-  scene.on('loaded', () => {
-    fetch('https://geo.datav.aliyun.com/areas_v3/bound/330000.json')
-      .then((res) => res.json())
-      .then((data) => {
-        const provincelayerTop = new PolygonLayer({
-          autoFit: true,
-        })
-          .source(data)
-          .size(1000)
-          .shape('extrude')
-          .size(10000)
-          .color('#0DCCFF')
-          //   .active({
-          //     color: 'rgb(100,230,255)',
-          //   })
-          .style({
-            heightfixed: true,
-            // pickLight: true,
-            mapTexture:
-              'https://mdn.alipayobjects.com/huamei_qa8qxu/afts/img/A*bm0eRKOCcNoAAAAAAAAAAAAADmJ7AQ/original',
-            // mapTexture:'https://mdn.alipayobjects.com/huamei_qa8qxu/afts/img/A*3UcORbAv_zEAAAAAAAAAAAAADmJ7AQ/original',
-            // raisingHeight: 10000,
-            opacity: 0.8,
-            // sidesurface: false,
-          });
-        scene.addLayer(provincelayerTop);
-
-        if (window['screenshot']) {
-          window['screenshot']();
-        }
-      });
-  });
-}
