+++ examples/demos_next/polygon/fill-linear.ts
@@ -1,16 +1,14 @@
-import { PolygonLayer, Scene } from '@antv/l7';
-import * as allMap from '@antv/l7-maps';
-import type { RenderDemoOptions } from '../../types';
+import { PolygonLayer } from '@antv/l7';
+import type { TestCase } from '../../types';
+import { CaseScene } from '../../utils';
 
-export function MapRender(options: RenderDemoOptions) {
-  const scene = new Scene({
-    id: 'map',
-    renderer: options.renderer,
-    map: new allMap[options.map]({
-      style: 'light',
+export const fillLinear: TestCase = async (options) => {
+  const scene = await CaseScene({
+    ...options,
+    mapConfig: {
       center: [121.434765, 31.256735],
       zoom: 14.83,
-    }),
+    },
   });
 
   const data = {
@@ -52,10 +50,7 @@ export function MapRender(options: RenderDemoOptions) {
       },
     });
 
-  scene.on('loaded', () => {
-    scene.addLayer(layer);
-    if (window['screenshot']) {
-      window['screenshot']();
-    }
-  });
-}
+  scene.addLayer(layer);
+
+  return scene;
+};
